Видалення сенсорів блоків(поки коментарієм) , заміна вирахування у якій імено сторіні відбулося зіткнення вирахуванням цього у методі WorldContactListner заміна у Block сетах контактів з блоками зліва і зправа, типу обєктів на AbstractGameObject

+++ core/src/com/daleondeveloper/Sprites/GameSensor.java
@@ -77,10 +77,19 @@ public class GameSensor extends AbstractDynamicObject {
     public void update(float deltaTime) {
 
         if(firstLineBlocks.size() > 11){
+            int blockCountInFirstLine = 0;
             for(Block block : firstLineBlocks){
-                block.delete();
+                if(block.isIdle()){
+                    blockCountInFirstLine++;
+                }
             }
-            playScreen.getHud().addScore(10);
+            if(blockCountInFirstLine > 11) {
+                for (Block block : firstLineBlocks) {
+                    block.delete();
+                }
+
+                playScreen.getHud().addScore(10);
+            }
         }
 
     }
